pyproject.toml
[project]
name = "tts-rag-workshop"
version = "0.1.0"
description = "虚拟销售 AI 系统 - RAG 实验开发项目"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "numpy>=1.20.0",
    "requests>=2.28.0",
    "jieba>=0.42.1",

    # Azure服务
    "azure-cognitiveservices-speech>=1.38.0",
    "azure-search-documents>=11.4.0",

    # 文档处理
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",

    # 音频处理
    "pyaudio>=0.2.14",

    # 数据处理
    "pandas>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=7.0.0",
    "pandas>=1.1.0",
    "tabulate>=0.8.0",
    "pytest>=6.0.0",
    "pytest-benchmark>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["experiment"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipython>=7.0.0",
    "pandas>=1.1.0",
    "tabulate>=0.8.0",
    "pytest>=6.0.0",
    "pytest-benchmark>=3.4.0",
]
